feat : ($CarNames, $Cars, $RacingChance) 일반 class 에서 record 클래스로 변경하고 iterator를 내부에 구현
세 클래스 모두 값이 변경되면 안되므로 record로 변경함으로 이득이 크다고 판단함
