+++ src/main/java/racingcar/model/entity/Cars.java
@@ -1,21 +1,17 @@
 package racingcar.model.entity;
 
+import racingcar.model.dto.CarNames;
 import racingcar.strategy.MoveStrategy;
 
 import java.util.Iterator;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.stream.StreamSupport;
 
 /**
  * 자동차들을 콜렉션으로 저장한 클래스
  */
-public class Cars implements Iterable<Car> {
-
-    private final List<Car> cars;
-
-    private Cars(List<Car> cars) {
-        this.cars = cars;
-    }
+public record Cars(List<Car> cars) implements Iterable<Car> {
 
     public static Cars getOfNamesAndStrategy(CarNames names, MoveStrategy strategy) {
         List<Car> elements = StreamSupport.stream(names.spliterator(), false)
@@ -26,15 +22,7 @@ public class Cars implements Iterable<Car> {
 
     @Override
     public Iterator<Car> iterator() {
-        return new CarsIterator(this);
-    }
-
-    protected int getLength() {
-        return cars.size();
-    }
-
-    protected Car getCarAt(int index) {
-        return cars.get(index);
+        return new CarsIterator();
     }
 
     public List<String> getMaxProgressCarsName() {
@@ -48,4 +36,27 @@ public class Cars implements Iterable<Car> {
                 .toList();
     }
 
+    private final class CarsIterator implements Iterator<Car> {
+
+        private int index;
+
+        private CarsIterator() {
+            this.index = 0;
+        }
+
+        @Override
+        public boolean hasNext() {
+            return index < cars.size();
+        }
+
+        @Override
+        public Car next() {
+            if (!hasNext()) {
+                throw new NoSuchElementException();
+            }
+            Car car = cars.get(index);
+            index++;
+            return car;
+        }
+    }
 }
