+++ src/test/java/racingcar/ExceptionMessageTest.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.model.entity.CarNames;
+import racingcar.model.dto.CarNames;
 import racingcar.model.entity.RacingChance;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
